Decoupled storage metadata from storage organisation
