+++ src/metadata.rs
@@ -164,6 +164,16 @@ impl StorageMetadata {
         serde_json::to_writer(&file, &self.metadata).expect("Failed to write");
     }
 
+    pub fn move_all(&mut self, moved_files: &[(u32, u32)]) {
+        for entry in self.metadata.iter_mut() {
+            for moved_file in moved_files.iter() {
+                if moved_file.0 == entry.id {
+                    entry.move_id(moved_file.1);
+                }
+            }
+        }
+    }
+
     fn name_with_id_predicate(index: u32, entry: io::Result<DirEntry>) -> bool {
         match entry {
             Ok(dir_entry) => {
